Revert "elementId int32_t -> int64_t"
This reverts commit bd248e1c345b29c763a62a9a3ecd20e51eaa4ac0.

+++ common/interface/include/i_accessible_ability_channel.h
@@ -44,7 +44,7 @@ public:
      * @return Return RET_OK if search elementInfo successfully, otherwise refer to the RetError for the failure.
      */
     virtual RetError SearchElementInfoByAccessibilityId(const int32_t accessibilityWindowId,
-        const int64_t elementId, const int32_t requestId, const sptr<IAccessibilityElementOperatorCallback> &callback,
+        const int32_t elementId, const int32_t requestId, const sptr<IAccessibilityElementOperatorCallback> &callback,
         const int32_t mode) = 0;
 
     /**
@@ -57,7 +57,7 @@ public:
      * @param callback  To transfer the element info to ASAC and it defined by ASAC.
      * @return Return RET_OK if search elementInfo successfully, otherwise refer to the RetError for the failure.
      */
-    virtual RetError SearchElementInfosByText(const int32_t accessibilityWindowId, const int64_t elementId,
+    virtual RetError SearchElementInfosByText(const int32_t accessibilityWindowId, const int32_t elementId,
         const std::string &text, const int32_t requestId,
         const sptr<IAccessibilityElementOperatorCallback> &callback) = 0;
 
@@ -71,7 +71,7 @@ public:
      * @param callback  To transfer the element info to ASAC and it defined by ASAC.
      * @return Return RET_OK if find elementInfo successfully, otherwise refer to the RetError for the failure.
      */
-    virtual RetError FindFocusedElementInfo(const int32_t accessibilityWindowId, const int64_t elementId,
+    virtual RetError FindFocusedElementInfo(const int32_t accessibilityWindowId, const int32_t elementId,
         const int32_t focusType, const int32_t requestId,
         const sptr<IAccessibilityElementOperatorCallback> &callback) = 0;
 
@@ -84,7 +84,7 @@ public:
      * @param callback  To transfer the element info to ASAC and it defined by ASAC.
      * @return Return RET_OK if search elementInfo successfully, otherwise refer to the RetError for the failure.
      */
-    virtual RetError FocusMoveSearch(const int32_t accessibilityWindowId, const int64_t elementId,
+    virtual RetError FocusMoveSearch(const int32_t accessibilityWindowId, const int32_t elementId,
         const int32_t direction, const int32_t requestId,
         const sptr<IAccessibilityElementOperatorCallback> &callback) = 0;
 
@@ -98,7 +98,7 @@ public:
      * @param callback  To transfer the node info to ASAC and it defined by ASAC.
      * @return Return RET_OK if perform action successfully, otherwise refer to the RetError for the failure.
      */
-    virtual RetError ExecuteAction(const int32_t accessibilityWindowId, const int64_t elementId, const int32_t action,
+    virtual RetError ExecuteAction(const int32_t accessibilityWindowId, const int32_t elementId, const int32_t action,
         const std::map<std::string, std::string> &actionArguments, const int32_t requestId,
         const sptr<IAccessibilityElementOperatorCallback> &callback) = 0;
