+++ frameworks/aafwk/test/mock/include/mock_accessible_ability_channel_proxy.h
@@ -27,18 +27,18 @@ public:
     virtual ~MockAccessibleAbilityChannelProxy() = default;
 
     MOCK_METHOD5(SearchElementInfoByAccessibilityId,
-        RetError(const int32_t accessibilityWindowId, const int64_t elementId, const int32_t requestId,
+        RetError(const int32_t accessibilityWindowId, const int32_t elementId, const int32_t requestId,
             const sptr<IAccessibilityElementOperatorCallback>& callback, const int32_t mode));
     MOCK_METHOD5(SearchElementInfosByText,
-        RetError(const int32_t accessibilityWindowId, const int64_t elementId, const std::string& text,
+        RetError(const int32_t accessibilityWindowId, const int32_t elementId, const std::string& text,
             const int32_t requestId, const sptr<IAccessibilityElementOperatorCallback>& callback));
     MOCK_METHOD5(FindFocusedElementInfo,
-        RetError(const int32_t accessibilityWindowId, const int64_t elementId, const int32_t focusType,
+        RetError(const int32_t accessibilityWindowId, const int32_t elementId, const int32_t focusType,
             const int32_t requestId, const sptr<IAccessibilityElementOperatorCallback>& callback));
-    MOCK_METHOD5(FocusMoveSearch, RetError(const int32_t accessibilityWindowId, const int64_t elementId,
+    MOCK_METHOD5(FocusMoveSearch, RetError(const int32_t accessibilityWindowId, const int32_t elementId,
         const int32_t direction, const int32_t requestId,
         const sptr<IAccessibilityElementOperatorCallback>& callback));
-    MOCK_METHOD6(ExecuteAction, RetError(const int32_t accessibilityWindowId, const int64_t elementId,
+    MOCK_METHOD6(ExecuteAction, RetError(const int32_t accessibilityWindowId, const int32_t elementId,
         const int32_t action, const std::map<std::string, std::string>& actionArguments, const int32_t requestId,
         const sptr<IAccessibilityElementOperatorCallback>& callback));
     MOCK_METHOD2(GetWindow, RetError(const int32_t windowId, AccessibilityWindowInfo &windowInfo));
